update MSToolkit from b41b33f commit on 1/30/19
git-svn-id: https://svn.code.sf.net/p/comet-ms/code@1285 299d2d02-8f3b-485e-a07e-71e8e5e16577

+++ trunk/comet-ms/MSToolkit/src/MSToolkit/MSObject.cpp
@@ -109,6 +109,6 @@ void MSObject::setHeader(const MSHeader& h){
 };
 
 int MSObject::size(){
-  return vSpectrum->size();
+  return (int)vSpectrum->size();
 };
